Update project dependencies
Replaced flask with aiohttp. This is because concurrency would be easier
to implement with aiohttp. As part of this change, a custom event loop
will be used. This will be uvloop.

On Google App Engine the default server is gunicorn and declaring it in
requirements.txt is not necessary. But to get aiohttp and uvloop to work
with gunicorn, custom conditions to be specified hence needing gunicorn
to be declared.

To simplify logging, loguru will be used.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>