Fix missing dependency
Setuptools is a dependency of gunicorn. Poetry considers setuptools, distribute
and pip to be unsafe packages and therefore doesn't include them when
generating the lock file (poetry.lock) This subsequently means the packages
aren't included when exporting to requirements.txt.

A workaround is to go to ~/.poetry/lib/poetry/puzzle/provider.py. In the
class Provider, there's a class variable called UNSAFE_PACKAGES which is
a set. Removing setuptools from that set will allow  poetry to add it to the
poetry.lock when the command "poetry lock" is called.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>