Update Makefile
Previously, the makefile would check for the protobuf output file in the
root directory instead of the src directory. This meant that the protobuf
rule would always execute even if the file was present.

Added rules to deploy the default service to Google App Engine. As part
of this, there's a rule to create the requirements.txt file that GAE needs.

Also, added a rule to help local deployment easier.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>