Improve type hinting of protobuf output files
Added a protoc plugin by Dropbox, mypy-protobuf, that generates type
hint stubs based on the .proto file used. Like the protobuf python output,
the stub file is not added to version control.

Also, a missing dependency for the protobuf package was added. This
package does the work of converting the protobuf output file info into
actual python Classes and types.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>