Update protobuf schema
In metadata, instead of having blue_team and red_team, repeated was
chosen as an alternative with the side being indicated within the respective
teams. This would make it easier to create and parse the metadata. As part
of this change, the Side enum was removed from Event as it was no longer
relevant to have it like so.

Since riot has an id parameter in the feed data, parameter is now used to
indicate a vod's id on it's specific platform. Id is now used for the same
purpose as in riot's data.

Added locale to vods as it may be useful in future.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>