Add script to handle api requests
To ensure that one aiohttp ClientSession object was used for multiple
requests, an application object would be stored as a class instance. The
application object would contain within it a ClientSession object to be
used in requests.

The initialisation of the clientSession object would occur during the
startup phase of the web server.

Signed-off-by: Victor Otieno <vickz84259@gmail.com>