Add ClientSession initialisation code
To take advantage of connection pools, one aiohttp.ClientSession object
will be used for the lifetime of the server application. To facilitate this,
the object is added to the application object before being added to the
State class as a class instance. Although use of a class variable is
discouraged by the developers of aiohttp, there was no other viable
solution in this case.

The changes to the handler configuration is only for local deployment.
On Google App Engine, the sink will be reverted back to sys.stderr

Signed-off-by: Victor Otieno <vickz84259@gmail.com>